update nav mobile
