+++ src/pages/_app.tsx
@@ -2,6 +2,7 @@ import "@/styles/globals.css";
 import type { AppProps } from "next/app";
 import Head from "next/head";
 import { ReactLenis } from "lenis/dist/lenis-react";
+import Script from "next/script";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
@@ -37,6 +38,20 @@ export default function App({ Component, pageProps }: AppProps) {
         <meta name="twitter:image" content="/meta/twitter-image.png" />
       </Head>
 
+      <Script
+        src="https://code.jquery.com/jquery-3.7.1.min.js"
+        strategy="afterInteractive"
+        onLoad={() => {
+          // Load Mailchimp code after jQuery is available
+          const script = document.createElement("script");
+          script.text = `(function($) {window.fnames = new Array(); window.ftypes = new Array();fnames[0]="EMAIL";ftypes[0]="email";fnames[1]="FNAME";ftypes[1]="text";fnames[2]="LNAME";ftypes[2]="text";fnames[3]="ADDRESS";ftypes[3]="address";fnames[4]="PHONE";ftypes[4]="phone";fnames[5]="BIRTHDAY";ftypes[5]="birthday";}(jQuery));var $mcj = jQuery.noConflict(true);`;
+          document.body.appendChild(script);
+        }}
+      />
+      <Script
+        src="//s3.amazonaws.com/downloads.mailchimp.com/js/mc-validate.js"
+        strategy="afterInteractive"
+      />
       <ReactLenis root>
         <Component {...pageProps} />
       </ReactLenis>
